refactor(runtime-tags): clear residue left by completed translator refactors
