+++ packages/runtime-tags/AGENTS.md
@@ -63,7 +63,7 @@ export default {
 
 - **`_name` exports** are runtime API called by generated code — public to codegen, not to app authors. Renames must update `callRuntime` call sites and `pureDOMFunctions`.
 - **`MARKO_DEBUG`** gates all validation, descriptive names, and detailed error messages (`if (MARKO_DEBUG) { ... }`); builds strip these. It is `true` in tests via the `~ts` register hook. Runtime error helpers live in `common/errors.ts`.
-- **`.debug.ts` pairs**: source imports the `.debug` module (e.g. `common/types.ts` imports `./constants/accessor-prop.debug`); the production build remaps `X.debug` → `X.ts`. Both files must export identical member names. There are nine pairs: seven accessor/enum modules under `src/common/constants/`, plus `common/accessor[.debug].ts` and `html/inlined-runtimes[.debug].ts`. Nothing enforces parity — each module's `Value` is `typeof import("./<itself>")` and `translator/util/get-accessor-char.ts` casts `as any` — so a member added to only one half type-checks and shows up as `undefined` in an optimize build.
+- **`.debug.ts` pairs**: source imports the `.debug` module (e.g. `common/types.ts` imports `./constants/accessor-prop.debug`); the production build remaps `X.debug` → `X.ts`. Both files must export identical member names. There are nine pairs: seven accessor/enum modules under `src/common/constants/`, plus `common/accessor[.debug].ts` and `html/inlined-runtimes[.debug].ts`. Nothing enforces parity — each module's `Value` is `typeof import("./<itself>")` and `translator/util/get-accessor-enums.ts` casts `as any` — so a member added to only one half type-checks and shows up as `undefined` in an optimize build.
 - **Optional feature enablement** (tree-shakable runtime API) has two patterns:
   - **`src/{dom,html}/**/*.feat.ts`** are compiler-injected side-effect modules for behavior a referenced import cannot keep alive (catch handling, controllable registration). The build emits them as extra entries of their runtime's bundle — a shared chunk keeps one state instance, so they import runtime internals directly. A feature body is direct registry/property assignments (plus at most one installer call, e.g. `installCatch`); the compiler emits it once per program via `importRuntimeFeature` (typed by `DOMRuntimeFeature`).
   - **Definition-site wrappers** gate behavior on a helper's own retention: `export const _if = /*@__PURE__*/ withBranches(...)`.
