+++ packages/runtime-tags/src/translator/core/let.ts
@@ -7,7 +7,7 @@ import {
 } from "@marko/compiler/babel-utils";
 
 import { assertNoBodyContent, assertNoSpreadAttrs } from "../util/assert";
-import { getAccessorPrefix } from "../util/get-accessor-char";
+import { getAccessorPrefix } from "../util/get-accessor-enums";
 import { isOutputDOM } from "../util/marko-config";
 import {
   BindingType,
